ai-governance-dashboard/email-agent/backend/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# Cache & Queue
redis>=5.0.1
celery>=5.3.6

# AI & LangGraph
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.3.0

# Google APIs
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# AWS Services
boto3>=1.34.34
botocore>=1.34.34

# Authentication
pyjwt[crypto]>=2.8.0
cryptography>=42.0.0
passlib>=1.7.4
python-jose[cryptography]>=3.3.0

# HTTP & WebSocket
httpx>=0.26.0
websockets>=12.0
python-multipart>=0.0.6

# Utilities
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0
python-dateutil>=2.8.2
pytz>=2024.1

# Monitoring & Logging
prometheus-client>=0.19.0
structlog>=24.1.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
hypothesis>=6.98.0
faker>=22.6.0
